FEAT: Implement signin/signout page for backoffice page. Backend notion graffix requests only available when signed in. Remove router in graffix-requests, exchanged for dropdown to switch between departments
