+++ src/pages/api/auth/[...nextauth].js
@@ -9,6 +9,7 @@ export const authOptions = {
       tenantId: process.env.AZURE_AD_TENANT_ID,
     }),
   ],
+  secret: process.env.NEXTAUTH_SECRET
 }
 
 export default NextAuth(authOptions)
