+++ src/pages/graffix/backoffice/[id].tsx
@@ -1,13 +1,12 @@
 import Head from 'next/head';
 import { Page } from 'modules';
-import { useRouter } from 'next/router';
+// import { useRouter } from 'next/router';
 
 import { FluidContainer, Typography } from 'components';
 import Link from 'next/link';
 
 export default function GraphicsRequests() {
-  const router = useRouter();
-  const { id } = router.query;
+  // const router = useRouter();
 
   return (
     <Page>
@@ -19,14 +18,14 @@ export default function GraphicsRequests() {
           Hi! This page has been moved to:
         </Typography>
         <Link
-          href={`/backoffice/graffix-requests/${id}`}
+          href={`/backoffice/graffix-requests`}
           style={{
             textDecoration: 'underline',
             marginTop: '1rem',
             display: 'block',
           }}
         >
-          https://www.calstatela.edu/backoffice/graffix-requests/{id}{' '}
+          https://www.calstatela.edu/backoffice/graffix-requests
         </Link>
       </FluidContainer>
     </Page>
